+++ core/lib/Thelia/Controller/BaseController.php
@@ -279,6 +279,42 @@ abstract class BaseController implements ControllerInterface
     }
 
     /**
+     * Setup the error context when an error occurs in a action method.
+     *
+     * @param string          $action        the action that caused the error (category modification, variable creation, currency update, etc.)
+     * @param BaseForm        $form          the form where the error occured, or null if no form was involved
+     * @param string          $error_message the error message
+     * @param \Exception|null $exception     the exception or null if no exception
+     */
+    protected function setupFormErrorContext(string $action, string $error_message, BaseForm $form = null, \Exception $exception = null): void
+    {
+        if ($error_message !== false) {
+            // Log the error message
+            Tlog::getInstance()->error(
+                $this->getTranslator()->trans(
+                    'Error during %action process : %error. Exception was %exc',
+                    [
+                        '%action' => $action,
+                        '%error' => $error_message,
+                        '%exc' => $exception != null ? $exception->getMessage() : 'no exception',
+                    ]
+                )
+            );
+
+            if ($form != null) {
+                // Mark the form as errored
+                $form->setErrorMessage($error_message);
+
+                // Pass it to the parser context
+                $this->getParserContext()->addForm($form);
+            }
+
+            // Pass the error message to the parser.
+            $this->getParserContext()->setGeneralError($error_message);
+        }
+    }
+
+    /**
      * @param int    $order_id
      * @param string $fileName
      * @param bool   $checkOrderStatus
